feat:added server fetching for movies
